Merge issueFreeRoutePass with refundRoutePass
This change is to make it easier for operators to perform work. Instead of refunding to routeCredits once, then issuing more routeCredits as compensation, it seems to make more sense to just issue the credits outright, especially since, for these use cases, there is no need to refund the ticket already - trip has already passed. An option to refund / void the ticket is added to the interface in case it is still needed.